minor adjustments - fix warnings - make variables static - eliminate non-generic check in HandleMenuAction

+++ Source/scrollrt.cpp
@@ -221,7 +221,7 @@ static void scrollrt_draw_cursor()
 	sgCursHgt -= sgCursY;
 	sgCursHgt++;
 
-	assert(sgCursWdt * sgCursHgt <= sizeof sgSaveBack);
+	assert((unsigned)(sgCursWdt * sgCursHgt) <= sizeof(sgSaveBack));
 	assert(gpBuffer != NULL);
 	dst = sgSaveBack;
 	src = &gpBuffer[SCREENXY(sgCursX, sgCursY)];
